reviso cada microservicio y verifico q anden los endpoint , ademas creo endpoint en postman todo perfecto solo quedo el administrador verificar y terminar

+++ TPEPARTE1/msViaje/src/main/java/com/TPE/msViaje/model/Viaje.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class Viaje {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private int idViaje;
+    private Long idViaje;
 
     private Long monopatinId;
     private LocalDateTime fechaInicio;
